+++ AlchemistCode/Powers/RerunPower.cs
@@ -1,0 +1,36 @@
+using System.Collections.Generic;
+using System.Linq;
+using MegaCrit.Sts2.Core.Combat;
+using MegaCrit.Sts2.Core.Commands;
+using MegaCrit.Sts2.Core.Entities.Creatures;
+using MegaCrit.Sts2.Core.Entities.Powers;
+using MegaCrit.Sts2.Core.GameActions.Multiplayer;
+using MegaCrit.Sts2.Core.Models;
+
+namespace Alchemist.AlchemistCode.Powers;
+
+// Base BurstPower without the Skill restriction: the next card of any type is played an extra
+// time, one stack per card, gone at the end of the turn
+public class RerunPower : AlchemistPower
+{
+    public override PowerType Type => PowerType.Buff;
+    public override PowerStackType StackType => PowerStackType.Counter;
+
+    public override int ModifyCardPlayCount(CardModel card, Creature? target, int playCount)
+    {
+        if (card.Owner.Creature != Owner) return playCount;
+        return playCount + 1;
+    }
+
+    public override async Task AfterModifyingCardPlayCount(CardModel card)
+    {
+        await PowerCmd.Decrement(this);
+    }
+
+    public override async Task AfterSideTurnEnd(PlayerChoiceContext choiceContext, CombatSide side,
+        IEnumerable<Creature> participants)
+    {
+        if (participants.Contains(Owner))
+            await PowerCmd.Remove(this);
+    }
+}