+++ AlchemistCode/Relics/SecondSkin.cs
@@ -15,7 +15,7 @@ public class SecondSkin : AlchemistRelic
 {
     private const int Block = 2;
 
-    public override RelicRarity Rarity => RelicRarity.Uncommon;
+    public override RelicRarity Rarity => RelicRarity.Common;
 
     protected override IEnumerable<IHoverTip> ExtraHoverTips =>
         new[] { HoverTipFactory.FromPower<PoisonPower>() };